+++ tests/Tl.Core.Tests/UniformRunRoutingTests.cs
@@ -1,7 +1,6 @@
 using System.Runtime.CompilerServices;
 using Xunit;
 
-using Tl.TestSupport;
 
 namespace Tl.Core.Tests;
 
@@ -49,7 +48,7 @@ public class UniformRunRoutingTests
         var baker = new Baker()
             .Track<RoutingTrack, RoutingClip>(new RoutingTrack(scale))
             .Clip(0, 0u, (uint)(duration * 6 / 10), new RoutingClip(1.25f))
-            .Clip(0, (uint)(duration * 6 / 10), (uint)duration, new RoutingClip(-0.5f));
+            .Clip(0, (uint)(duration * 6 / 10), duration, new RoutingClip(-0.5f));
         if (looping) baker.Looping();
         return baker.Bake();
     }
