+++ tools/Tl.Gen.Tlb/FastNumber.cs
@@ -1,4 +1,3 @@
-using System;
 
 namespace Tl.Gen.Tlb;
 
@@ -288,7 +287,7 @@ internal static class FastNumber
         var k = -exponent;
         if (k <= 10 && mantissa <= 0xFFFFFF)
         {
-            var divided = (float)((double)mantissa / Pow10Double[k]);
+            var divided = (float)(mantissa / Pow10Double[k]);
             return float.IsInfinity(divided) ? float.NaN : (negative ? -divided : divided);
         }
         return float.NaN;
@@ -304,9 +303,9 @@ internal static class FastNumber
             return double.NaN;
         double value;
         if (exponent >= 0)
-            value = (double)mantissa * Pow10Double[exponent];
+            value = mantissa * Pow10Double[exponent];
         else
-            value = (double)mantissa / Pow10Double[-exponent];
+            value = mantissa / Pow10Double[-exponent];
         if (double.IsInfinity(value))
             return double.NaN;
         return negative ? -value : value;
